! Fixed energy display

+++ src/plantplayer.cpp
@@ -162,7 +162,7 @@ void PlantPlayer::DrawEnergy() {
     if (tmpFlower->energy_cost > energy)
       color = CL_Colorf::red;
     default_font.draw_text(*gc, CL_Pointf(10, 30),
-                           cl_format("Energy: %1 (%2)", energy,
+                           cl_format("Energy: %1 (%2)", (int)energy,
                                      tmpFlower->energy_cost), color);
     break;
   }
